Uses send notification from use case everywhere
