+++ apps/judicial-system/web/src/utils/hooks/use-case/sendNotificationGql.ts
@@ -1,0 +1,9 @@
+import { gql } from '@apollo/client'
+
+export const SendNotificationMutation = gql`
+  mutation SendNotificationMutation($input: SendNotificationInput!) {
+    sendNotification(input: $input) {
+      notificationSent
+    }
+  }
+`
